Add styling to info section 2nd container
Added styling to info-container-two, .box, and all grouped boxes and
text. Made container-two into a flexbox. The box container was given
a height, width, rounded corners, and a border. The grouped boxes and
text were made into flexbox, put into a column direction, and wrapped.
Giving the grouped boxes a set width means that the text doesn't go past
the width of the boxes. Text was styled and boxes were given a margin to
spread them apart.