Add styling to the message in quote-section
Turned 'container-quote' into a flexbox for easier manipulation of
content. Wrapped the text so that the sun tzu item would position
below and off to the side. Added a width so that the text would wrap
centrally. Also introduced some styling to the text like size and
style. Sun Tzu P was given an id selector so that it could be
separately styled to the main P. Changes were made to the font weight,
size, and style for Sun Tzu.