Add all html for the cta section
Added the block of html for the cta section of the landing page.
Consists of several elements and containers. Added cta-section
with cta-banner nested within for centering content and to create
flexbox for further nested containers. Added cta-text and cta-button
into cta-banner to contain and position the content with flexbox.
Within cta-text was added the text portion of the section under
two P elements with id selectors for individual styling.
Within cta-button was added a button with class of sign-up-2 for styling
 and for differentiation to earlier button.